refactor(ViewHolder): ViewHolder의 bind 방법 수정
