+++ app/src/main/java/woowacourse/movie/viewholder/AdViewHolder.kt
@@ -1,13 +1,15 @@
-package woowacourse.movie.movieList
+package woowacourse.movie.viewholder
 
 import android.view.View
 import android.widget.ImageView
 import woowacourse.movie.R
 import woowacourse.movie.uimodel.MovieModelUi
 class AdViewHolder(
-    private val view: View,
+    view: View,
 ) {
+    private val adImageView = view.findViewById<ImageView>(R.id.ad_image_view)
+
     fun bind(adUi: MovieModelUi.AdUi) {
-        view.findViewById<ImageView>(R.id.ad_image_view).setImageResource(adUi.addPoster)
+        adImageView.setImageResource(adUi.addPoster)
     }
 }
